mes: ReadMesChar returns u16, drops redundant conversions in addString (92.19->92.33)

+++ src/mes.cpp
@@ -94,7 +94,7 @@ static inline int ReadTagS8(char** text)
 	return (int)ReadTagByte(text);
 }
 
-static inline int ReadMesChar(char** text)
+static inline unsigned short ReadMesChar(char** text)
 {
 	unsigned char* p = (unsigned char*)*text;
 	*text = (char*)(p + 1);
